# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Web Development Course API
  version: 0.1.0
servers:
  - url: http://localhost:7001
paths:
  /api/health:
    get:
      operationId: getHealth
      summary: 检查 API 状态
      responses:
        "200":
          description: API 正常
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  /api/courses:
    get:
      operationId: listCourses
      summary: 获取课程列表
      responses:
        "200":
          description: 课程列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      summary: 创建课程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourse"
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Course"
        "400":
          description: 请求体不符合课程数据规则

  # ---- Matches ----

  /api/matches:
    get:
      operationId: listMatches
      summary: 获取赛程列表
      parameters:
        - name: league
          in: query
          schema:
            $ref: "#/components/schemas/League"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MatchStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 赛程列表
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MatchSummary"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: 参数无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/matches/{matchId}:
    get:
      operationId: getMatch
      summary: 获取单场比赛详情
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 比赛详情
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/MatchDetail"
        "404":
          description: 比赛不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMatchStatus
      summary: 更新比赛状态（管理员）
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/MatchStatus"
      responses:
        "200":
          description: 状态更新成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/MatchDetail"
        "400":
          description: 非法状态转换
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 非管理员
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 比赛不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Teams ----

  /api/teams:
    get:
      operationId: listTeams
      summary: 获取球队列表
      parameters:
        - name: league
          in: query
          schema:
            $ref: "#/components/schemas/League"
      responses:
        "200":
          description: 球队列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
        "400":
          description: 参数无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/teams/{teamId}:
    get:
      operationId: getTeam
      summary: 获取单支球队详情
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 球队详情
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Team"
        "404":
          description: 球队不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Standings ----

  /api/standings:
    get:
      operationId: getStandings
      summary: 获取积分榜
      parameters:
        - name: league
          in: query
          schema:
            $ref: "#/components/schemas/League"
      responses:
        "200":
          description: 积分榜
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Standing"
        "400":
          description: 参数无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Bracket ----

  /api/bracket:
    get:
      operationId: getBracket
      summary: 获取淘汰赛对阵图
      parameters:
        - name: league
          in: query
          schema:
            $ref: "#/components/schemas/League"
      responses:
        "200":
          description: 淘汰赛对阵
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bracket"
        "400":
          description: 参数无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Predictions ----

  /api/predictions:
    post:
      operationId: createPrediction
      summary: 提交比分预测
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrediction"
      responses:
        "201":
          description: 预测创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Prediction"
        "200":
          description: 预测更新成功（已存在预测）
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Prediction"
        "400":
          description: 参数验证失败或比赛已开始
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listMyPredictions
      summary: 查看当前用户的预测列表
      responses:
        "200":
          description: 预测列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PredictionWithMatch"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/predictions/{predictionId}:
    put:
      operationId: updatePrediction
      summary: 修改比分预测（仅开赛前）
      parameters:
        - name: predictionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrediction"
      responses:
        "200":
          description: 预测更新成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Prediction"
        "400":
          description: 参数验证失败或比赛已开始
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 只能修改自己的预测
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 预测不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Match Results ----

  /api/matches/{matchId}/results:
    post:
      operationId: enterResult
      summary: 录入比赛结果（管理员）
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnterResult"
      responses:
        "201":
          description: 结果录入成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/MatchResult"
        "400":
          description: 参数无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 非管理员
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 比赛不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 结果已录入
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Favorites ----

  /api/favorites:
    post:
      operationId: addFavorite
      summary: 添加收藏
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFavorite"
      responses:
        "201":
          description: 收藏成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Favorite"
        "400":
          description: 参数无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 已收藏
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listMyFavorites
      summary: 查看当前用户的收藏列表
      responses:
        "200":
          description: 收藏列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Favorite"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/favorites/{favoriteId}:
    delete:
      operationId: removeFavorite
      summary: 取消收藏
      parameters:
        - name: favoriteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 取消成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 只能操作自己的收藏
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 收藏不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Comments ----

  /api/matches/{matchId}/comments:
    post:
      operationId: createComment
      summary: 发表评论
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComment"
      responses:
        "201":
          description: 评论成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Comment"
        "400":
          description: 内容无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 比赛不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listComments
      summary: 获取评论列表
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 评论列表
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

  /api/comments/{commentId}:
    delete:
      operationId: deleteComment
      summary: 删除评论（作者或管理员）
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Comment"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 只能删除自己的评论
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 评论不存在或已删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    # --- existing ---
    Health:
      type: object
      required: [status, service, timestamp]
      properties:
        status:
          const: ok
        service:
          type: string
        timestamp:
          type: string
          format: date-time
    Course:
      type: object
      required: [id, title, description, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateCourse:
      type: object
      additionalProperties: false
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          minLength: 2
          maxLength: 500

    # --- common ---
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 人类可读的错误信息

    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    League:
      type: string
      enum: [worldcup, spl]
      description: 联赛

    MatchStatus:
      type: string
      enum: [scheduled, live, finished, postponed, cancelled]
      description: 比赛状态

    MatchStage:
      type: string
      enum: [group, round16, quarter, semi, final]
      description: 比赛阶段

    FavoriteType:
      type: string
      enum: [match, team]
      description: 收藏类型

    # --- Match ---
    Match:
      type: object
      required:
        [id, homeTeamId, awayTeamId, kickoffTime, venue, league, status, stage, createdAt, updatedAt]
      properties:
        id:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        kickoffTime:
          type: string
          format: date-time
        venue:
          type: string
        league:
          $ref: "#/components/schemas/League"
        status:
          $ref: "#/components/schemas/MatchStatus"
        stage:
          $ref: "#/components/schemas/MatchStage"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TeamBrief:
      type: object
      required: [id, name, logoUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string

    MatchSummary:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          required: [homeTeam, awayTeam]
          properties:
            homeTeam:
              $ref: "#/components/schemas/TeamBrief"
            awayTeam:
              $ref: "#/components/schemas/TeamBrief"

    MatchDetail:
      allOf:
        - $ref: "#/components/schemas/MatchSummary"
        - type: object
          required: [commentCount]
          properties:
            commentCount:
              type: integer
              description: 关联评论数

    # --- Team ---
    Team:
      type: object
      required: [id, name, logoUrl, league, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
        league:
          $ref: "#/components/schemas/League"
        createdAt:
          type: string
          format: date-time

    # --- Standing ---
    Standing:
      type: object
      required:
        [id, teamId, teamName, league, played, won, drawn, lost, goalsFor, goalsAgainst, goalDifference, points]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        league:
          $ref: "#/components/schemas/League"
        played:
          type: integer
        won:
          type: integer
        drawn:
          type: integer
        lost:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        goalDifference:
          type: integer
        points:
          type: integer

    # --- Bracket ---
    Bracket:
      type: object
      required: [id, league, stage, position]
      properties:
        id:
          type: integer
        league:
          $ref: "#/components/schemas/League"
        stage:
          $ref: "#/components/schemas/MatchStage"
        matchId:
          type: integer
          nullable: true
        homeTeamId:
          type: integer
          nullable: true
        awayTeamId:
          type: integer
          nullable: true
        homeTeamName:
          type: string
          nullable: true
        awayTeamName:
          type: string
          nullable: true
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        parentNodeId:
          type: integer
          nullable: true
        position:
          type: string
          enum: [left, right]

    # --- Prediction ---
    CreatePrediction:
      type: object
      additionalProperties: false
      required: [matchId, homeScore, awayScore]
      properties:
        matchId:
          type: integer
        homeScore:
          type: integer
          minimum: 0
        awayScore:
          type: integer
          minimum: 0

    Prediction:
      type: object
      required: [id, userId, matchId, homeScore, awayScore, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        matchId:
          type: integer
        homeScore:
          type: integer
        awayScore:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    MatchBrief:
      type: object
      required: [id, homeTeamName, awayTeamName, kickoffTime, status, league]
      properties:
        id:
          type: integer
        homeTeamName:
          type: string
        awayTeamName:
          type: string
        kickoffTime:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/MatchStatus"
        league:
          $ref: "#/components/schemas/League"

    PredictionWithMatch:
      allOf:
        - $ref: "#/components/schemas/Prediction"
        - type: object
          required: [match]
          properties:
            match:
              $ref: "#/components/schemas/MatchBrief"

    # --- MatchResult ---
    EnterResult:
      type: object
      additionalProperties: false
      required: [homeScore, awayScore]
      properties:
        homeScore:
          type: integer
          minimum: 0
        awayScore:
          type: integer
          minimum: 0

    MatchResult:
      type: object
      required: [id, matchId, homeScore, awayScore, enteredBy, createdAt]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        homeScore:
          type: integer
        awayScore:
          type: integer
        enteredBy:
          type: string
        createdAt:
          type: string
          format: date-time

    # --- Favorite ---
    CreateFavorite:
      type: object
      additionalProperties: false
      required: [type, targetId]
      properties:
        type:
          $ref: "#/components/schemas/FavoriteType"
        targetId:
          type: integer

    Favorite:
      type: object
      required: [id, userId, type, targetId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          $ref: "#/components/schemas/FavoriteType"
        targetId:
          type: integer
        createdAt:
          type: string
          format: date-time

    # --- Comment ---
    CreateComment:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500

    Comment:
      type: object
      required: [id, userId, matchId, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        matchId:
          type: integer
        content:
          type: string
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time